pyproject.toml
[project]
name = "onc-data-wrangler-plugin"
version = "1.0.0"
description = "Self-contained Claude Code plugin for oncology data wrangling, extraction, database building, querying, and paper reproduction."
requires-python = ">=3.13"
dependencies = [
    "anthropic[vertex]>=0.42.0",
    "google-genai>=1.0.0",
    "duckdb>=1.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
    "rapidfuzz>=3.0.0",
    "sqlglot>=25.0.0",
    "lifelines",
    "statsmodels",
    "scikit-learn",
    "tqdm>=4.60.0",
    "scanpy>=1.12.1",
    "scikit-misc>=0.5.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/onc_wrangler"]
